Replace defaults
